Required Dynamic Libraries
Some libraries are loaded statically like core, but events and notes are required
to be instanciated. To be able to instanciate the objects for the api, they needed
to be on the core object, and needed the added security the API has so the
functions can be used as plugins as they take requests from the client at a core or
background level.

This was achieved by adding plugable library lists for core-level required libraries
that need created on-demand at runtime